requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
